+++ README.md
@@ -20,3 +20,7 @@ python app.py
 
 The home page at `http://localhost:5000/` lets you save tickers and search existing ones.
 Click any saved ticker to view the interactive chart at `/stock/<ticker>`.
+
+## Login
+
+The application now supports user accounts. Visit `/register` to create an account and `/login` to sign in. Once logged in, you can add and view saved tickers. Use `/logout` to end the session.
